cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autumn-cli"
version = "0.1.0"
description = "Command-line interface for AutumnWeb - time tracking and project management"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "AutumnWeb Contributors"}
]
keywords = ["time-tracking", "cli", "productivity", "project-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "tabulate>=0.9.0",
    "python-dateutil>=2.8.0",
    "wordcloud>=1.9.0",
    "rich>=13.0.0",
]

[project.scripts]
autumn = "autumn_cli.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/AutumnWeb"
Documentation = "https://github.com/yourusername/AutumnWeb/blob/main/cli/README.md"
Repository = "https://github.com/yourusername/AutumnWeb"

[tool.setuptools]
packages = ["autumn_cli", "autumn_cli.commands", "autumn_cli.utils"]

[tool.setuptools.package-data]
autumn_cli = ["py.typed"]
